tests: New test data for outdatad completions
