+++ database/data/08_leastcostchimps.csv
@@ -137,3 +137,521 @@
 137	999999999
 138	999999999
 139	139000
+708944	24
+18550	51
+113345	34
+74919	37
+402507	53
+170595	81
+839355	50
+180386	0
+210396	77
+358725	26
+22873	35
+414714	46
+971859	16
+476559	88
+956199	18
+814173	37
+535757	3
+314822	97
+150800	7
+883719	69
+91670	75
+889377	93
+257445	40
+795006	92
+140539	85
+125456	29
+626990	18
+460428	41
+227260	65
+827454	17
+717522	60
+615480	0
+541165	45
+536517	79
+666695	61
+276697	16
+563302	20
+399500	83
+165946	90
+67895	53
+630175	38
+980406	25
+977248	85
+284031	14
+31507	4
+370379	100
+147982	23
+742404	29
+111088	65
+829310	5
+993346	100
+18312	35
+80733	74
+525962	89
+712882	19
+129006	73
+275401	67
+202322	45
+290402	68
+606304	12
+826862	98
+370205	36
+269297	68
+797497	94
+669001	1
+65099	74
+824531	1
+794445	58
+338046	96
+952876	55
+315313	40
+903393	8
+940453	65
+625928	55
+902208	65
+714392	6
+640059	80
+412778	17
+769445	97
+100401	74
+990955	28
+616861	50
+297790	74
+967430	64
+517457	78
+255363	58
+621718	17
+781740	85
+571341	86
+120522	88
+817582	34
+436789	6
+390443	73
+738346	78
+539151	90
+340709	27
+424869	85
+354281	51
+661860	94
+350016	68
+191982	40
+520259	10
+11210	26
+373520	42
+895923	37
+687433	64
+220083	52
+732306	22
+330502	68
+338711	11
+246805	34
+740378	19
+580746	57
+27636	8
+251071	99
+405501	27
+785947	81
+141676	72
+70977	21
+907428	60
+101880	26
+106964	82
+704909	42
+89978	62
+529542	94
+124841	80
+59173	78
+545803	61
+802784	21
+523570	12
+655564	79
+851177	40
+433140	77
+781369	48
+778906	28
+577960	83
+416276	72
+413475	7
+202241	17
+318157	89
+707382	37
+463400	25
+616181	89
+33607	46
+242429	61
+596274	34
+967710	38
+242953	93
+507963	25
+165928	51
+806841	71
+896121	46
+910444	10
+710694	45
+625007	21
+165940	77
+719943	5
+672854	17
+70812	40
+764255	44
+911727	75
+164551	13
+855054	19
+66472	40
+717597	44
+488553	79
+894031	32
+834667	52
+915739	49
+465496	22
+775425	51
+408819	43
+664617	26
+891167	72
+395170	9
+868825	7
+59648	53
+170584	98
+592051	59
+803695	71
+166761	100
+139956	39
+148364	62
+683574	77
+251142	42
+714595	71
+661340	66
+835372	44
+969370	5
+496080	34
+35991	64
+404747	64
+35873	11
+856125	26
+487862	8
+294758	69
+360821	19
+141118	17
+300861	30
+440920	74
+61179	65
+756843	41
+591136	74
+296593	8
+418261	89
+904191	39
+913124	39
+857358	89
+594452	81
+113873	84
+815255	90
+58998	34
+317654	52
+743170	36
+134264	66
+10925	70
+410213	32
+817374	56
+88095	59
+430142	95
+914516	69
+935222	81
+333734	34
+738508	87
+679748	4
+846216	74
+988700	66
+691541	2
+611773	6
+584354	66
+638489	74
+58673	0
+706938	24
+881095	68
+247554	71
+919409	51
+916332	39
+204238	38
+364596	78
+505836	3
+757445	49
+68760	47
+606323	56
+20118	99
+673917	26
+220598	24
+448989	38
+119914	38
+188643	61
+229237	41
+262948	18
+981114	52
+338155	14
+919074	9
+323485	72
+528598	46
+283088	85
+210008	57
+968332	17
+805536	90
+287367	67
+338860	39
+222184	4
+252696	85
+13734	33
+392110	5
+871041	82
+589340	46
+121129	83
+371371	18
+706205	6
+697901	13
+809038	63
+952777	57
+216907	16
+487327	7
+811666	0
+447371	32
+926483	66
+122192	15
+180377	19
+906816	32
+114505	49
+430458	53
+164299	40
+216312	95
+107961	63
+607936	35
+188529	33
+572608	66
+425611	98
+850880	82
+643503	1
+321112	9
+240383	25
+97352	13
+624309	66
+164049	22
+465130	82
+192179	6
+424542	49
+913314	28
+758438	100
+885843	95
+163334	53
+892921	78
+97319	37
+418102	67
+273852	9
+398799	47
+555674	47
+168443	37
+166215	48
+612683	40
+315232	36
+168400	22
+719360	62
+929025	95
+830624	100
+924029	79
+765020	26
+503250	0
+794992	35
+210544	54
+928113	67
+326726	39
+608939	34
+97388	72
+21417	41
+691870	63
+511520	56
+132208	26
+40709	44
+796034	57
+849818	62
+356702	62
+130548	28
+779480	71
+515202	97
+89919	2
+252888	87
+805288	40
+853882	84
+188305	5
+562984	66
+800998	30
+363511	78
+855333	34
+690438	45
+46223	36
+48978	59
+626380	1
+615751	74
+588317	93
+432389	62
+131038	41
+183509	4
+720330	69
+738107	54
+504167	55
+13519	56
+876962	38
+952672	22
+46550	72
+958470	3
+711582	51
+867867	97
+785588	30
+821336	1
+494422	37
+963816	12
+631789	98
+29236	89
+914797	37
+466821	69
+464674	88
+437176	28
+424034	97
+244356	75
+871484	87
+281069	48
+789726	65
+419061	75
+471298	18
+432248	42
+80673	5
+502805	62
+879388	75
+743653	62
+280274	41
+776024	87
+258578	94
+605672	80
+858494	32
+469094	40
+962364	94
+381073	77
+199668	24
+526038	98
+903947	69
+636747	76
+168388	33
+643859	35
+417031	71
+415094	81
+935188	4
+576786	40
+871126	57
+537213	72
+209278	10
+661709	93
+707431	11
+966891	22
+159432	21
+144330	24
+895814	24
+641645	33
+370036	40
+796867	2
+453544	47
+574700	32
+710528	33
+866736	97
+252948	47
+906865	64
+732766	33
+148550	74
+207766	6
+351903	74
+446878	59
+110881	43
+635600	52
+600560	99
+735175	8
+536982	25
+161926	80
+563387	1
+772217	63
+21882	28
+798368	19
+290688	2
+474475	67
+800867	23
+307137	68
+860104	40
+302679	60
+588186	15
+981901	19
+670121	72
+72615	90
+292338	57
+829397	8
+186366	24
+842607	99
+595798	80
+393941	96
+853852	54
+279699	34
+153025	41
+139661	57
+129732	72
+616840	81
+378652	61
+592048	80
+998455	60
+270054	1
+725335	60
+690699	87
+755354	36
+367456	51
+984821	70
+848523	24
+522828	74
+662441	8
+418919	28
+893876	99
+480482	79
+671731	8
+696120	15
+30563	61
+14788	28
+770766	41
+488250	19
+222659	96
+279827	90
+166389	97
+210920	90
+977182	44
+850248	41
+447632	85
+449776	53
+910290	23
+499161	22
+982734	62
+111589	2
+931676	80
+525956	90
+194003	12
+396051	52
+541478	66
+636772	85
+325399	42
+623140	82
+136320	61
+101128	26
+158974	8
+514015	58
+457418	53
+919777	17
+468197	89
+930796	91
+273276	61
+210502	16
+289246	30
+332224	10
+286216	99
+216468	45
+153040	54
