+++ database/data/data_models/MapKey.py
@@ -1,0 +1,6 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class MapKey:
+    code: str
