+++ database/patches/2025_05_02_11_09_00-drop_new_version.psql
@@ -1,0 +1,1 @@
+SELECT * FROM maps LIMIT 1;
